Implement some more logic regarding memory bank controllers (1 and 5) (WIP)

+++ Core/include/Cartridge/MemoryBankController.hpp
@@ -24,7 +24,7 @@ namespace ggb
 		MC_INVALID, // Should be the last entry
 	};
 
-	uint16_t convertRawAddressToBankAddress(uint16_t address, int romBankNumber);
+	int convertRawAddressToBankAddress(uint16_t address, int romBankNumber);
 	MBCTYPE getMBCType(const std::vector<uint8_t>& cartRidgeData);
 
 	class MemoryBankController
